omniresearch/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
celery[redis]==5.4.0
anthropic==0.36.0
langgraph==0.2.0
langchain-core==0.3.0
sentence-transformers==3.3.0
transformers==4.46.0
torch==2.2.0
chromadb==0.5.23
httpx==0.27.0
arxiv==2.1.3
biopython==1.84
praw==7.8.1
youtube-transcript-api==0.6.3
yt-dlp==2024.11.18
openai-whisper==20240930
feedparser==6.0.11
pandas==2.2.0
reportlab==4.2.5
openpyxl==3.1.5
bertopic==0.16.4
pydantic-settings==2.6.0
redis==5.2.0
pytest==8.3.0
pytest-asyncio==0.24.0
